requirements.txt
amqp==5.2.0
APScheduler==3.10.4
asgiref==3.7.2
async-timeout==4.0.3
billiard==4.2.0
celery==5.3.6
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cryptography==41.0.5
defusedxml==0.8.0rc2
Deprecated==1.2.14
dj-rest-auth==5.0.2
Django==4.2.5
django-allauth==0.58.2
django-braces==1.15.0
django-cors-headers==4.2.0
django-cron==0.6.0
django-oauth-toolkit==2.3.0
django-rest-auth==0.9.5
django-rest-framework-social-oauth2==1.1.0
djangorestframework==3.14.0
drf-social-oauth2==2.1.3
ecdsa==0.18.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
jwcrypto==1.5.0
kombu==5.3.4
MarkupSafe==2.1.3
# mysqlclient==2.2.0
pymysql
oauthlib==3.2.2
Pillow==10.0.0
prompt-toolkit==3.0.41
pyasn1==0.5.0
pycparser==2.21
PyJWT==2.8.0
python-dateutil==2.8.2
python-decouple==3.8
python-jose==3.3.0
python3-openid==3.2.0
pytz==2023.3.post1
redis==5.0.1
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
schedule==1.2.1
six==1.16.0
social-auth-app-django==5.4.0
social-auth-core==4.5.0
sqlparse==0.4.4
typing_extensions==4.7.1
tzdata==2023.3
tzlocal==5.2
urllib3==2.0.7
vine==5.1.0
wcwidth==0.2.12
whitenoise==6.6.0
wrapt==1.16.0
